+++ Tests/MapboxMapsTests/Camera/Random/AnimationOwner+Random.swift
@@ -1,0 +1,7 @@
+import MapboxMaps
+
+extension AnimationOwner {
+    static func random() -> Self {
+        return AnimationOwner(rawValue: .randomASCII(withLength: .random(in: 0...10)))
+    }
+}